backend/requirement.txt
fastapi
uvicorn
python-multipart
pytesseract
opencv-python
Pillow
llama-index
openai
python-dotenv
